+++ apps/console/src/app.tsx
@@ -18,30 +18,30 @@
 
 import { BasicUserInfo, DecodedIDTokenPayload, useAuthContext } from "@asgardeo/auth-react";
 import { AccessControlProvider, AllFeatureInterface, FeatureGateInterface } from "@wso2is/access-control";
+import useAuthorization from "@wso2is/admin.authorization.v1/hooks/use-authorization";
+import { EventPublisher, PreLoader } from "@wso2is/admin.core.v1";
+import { ProtectedRoute } from "@wso2is/admin.core.v1/components";
+import CommonFeatureProviders from "@wso2is/admin.core.v1/components/common-feature-provider";
+import { Config, DocumentationLinks, getBaseRoutes } from "@wso2is/admin.core.v1/configs";
+import { AppConstants } from "@wso2is/admin.core.v1/constants";
+import { history } from "@wso2is/admin.core.v1/helpers";
+import useResourceEndpoints from "@wso2is/admin.core.v1/hooks/use-resource-endpoints";
+import {
+    ConfigReducerStateInterface,
+    DocumentationLinksInterface,
+    FeatureConfigInterface,
+    ServiceResourceEndpointsInterface
+} from "@wso2is/admin.core.v1/models";
+import { AppState } from "@wso2is/admin.core.v1/store";
+import { commonConfig } from "@wso2is/admin.extensions.v1";
+import { useGetAllFeatures } from "@wso2is/admin.extensions.v1/components/feature-gate/api/feature-gate";
+import { featureGateConfig } from "@wso2is/admin.extensions.v1/configs/feature-gate";
 import { AppConstants as CommonAppConstants } from "@wso2is/core/constants";
 import { IdentityAppsApiException } from "@wso2is/core/exceptions";
 import { CommonHelpers, isPortalAccessGranted } from "@wso2is/core/helpers";
 import { RouteInterface, StorageIdentityAppsSettingsInterface, emptyIdentityAppsSettings } from "@wso2is/core/models";
 import { setI18nConfigs, setServiceResourceEndpoints } from "@wso2is/core/store";
 import { AuthenticateUtils, LocalStorageUtils } from "@wso2is/core/utils";
-import useAuthorization from "@wso2is/features/admin.authorization.v1/hooks/use-authorization";
-import { EventPublisher, PreLoader } from "@wso2is/features/admin.core.v1";
-import { ProtectedRoute } from "@wso2is/features/admin.core.v1/components";
-import CommonFeatureProviders from "@wso2is/features/admin.core.v1/components/common-feature-provider";
-import { Config, DocumentationLinks, getBaseRoutes } from "@wso2is/features/admin.core.v1/configs";
-import { AppConstants } from "@wso2is/features/admin.core.v1/constants";
-import { history } from "@wso2is/features/admin.core.v1/helpers";
-import useResourceEndpoints from "@wso2is/features/admin.core.v1/hooks/use-resource-endpoints";
-import {
-    ConfigReducerStateInterface,
-    DocumentationLinksInterface,
-    FeatureConfigInterface,
-    ServiceResourceEndpointsInterface
-} from "@wso2is/features/admin.core.v1/models";
-import { AppState } from "@wso2is/features/admin.core.v1/store";
-import { commonConfig } from "@wso2is/features/admin.extensions.v1";
-import { useGetAllFeatures } from "@wso2is/features/admin.extensions.v1/components/feature-gate/api/feature-gate";
-import { featureGateConfig } from "@wso2is/features/admin.extensions.v1/configs/feature-gate";
 import { I18nModuleOptionsInterface } from "@wso2is/i18n";
 import {
     ChunkErrorModal,