+++ features/admin.application-roles.v1/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @wso2is/admin.application-roles.v1
 
+## 2.20.3
+
+### Patch Changes
+
+- Updated dependencies [[`de4201d75c779ee49c5bea1a63955f88f4b6b966`](https://github.com/wso2/identity-apps/commit/de4201d75c779ee49c5bea1a63955f88f4b6b966)]:
+  - @wso2is/admin.identity-providers.v1@2.20.3
+  - @wso2is/admin.applications.v1@2.20.3
+  - @wso2is/admin.connections.v1@2.20.3
+  - @wso2is/admin.extensions.v1@2.20.3
+  - @wso2is/admin.userstores.v1@2.20.3
+  - @wso2is/admin.groups.v1@2.20.3
+  - @wso2is/admin.core.v1@2.20.3
+
 ## 2.20.2
 
 ### Patch Changes